requirements-1.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
openai==1.75.0
chromadb==1.0.5
pypdf==5.4.0
python-docx==1.1.2
pydantic-settings==2.8.1
langchain-chroma==0.2.3
langchain-community==0.3.21
sentence-transformers==4.1.0
